# Conflicts:
#	Voting/pom.xml
#	Voting/target/classes/com/Voting/Voting/config/CassandraConfig.class
#	user-service/Dockerfile

+++ user-service/src/main/java/com/SWAFinalProject/userservice/controller/UserController.java
@@ -1,16 +1,15 @@
 package com.SWAFinalProject.userservice.controller;
 
+import com.SWAFinalProject.userservice.helpers.JwtHelper;
 import com.SWAFinalProject.userservice.entity.User;
+import com.SWAFinalProject.userservice.exception.NoContentFoundException;
+import com.SWAFinalProject.userservice.exception.UnAuthorizedException;
 import com.SWAFinalProject.userservice.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletRequest;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 
 @RestController
@@ -18,6 +17,8 @@ import java.util.UUID;
 public class UserController {
     @Autowired
     private UserService userService;
+
+    JwtHelper jwt = new JwtHelper();
 
 
     @PostMapping
@@ -40,15 +41,34 @@ public class UserController {
 
     @PutMapping("/{id}")
     public ResponseEntity<?> update(@PathVariable UUID id, @RequestBody User user, @RequestHeader("Authorization") String token) throws Exception {
-        userService.update(user, id, token);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (jwt.validateToken(token) && jwt.getUserFromToken(token).getUserId() != null) {
+            var allowedUser = userService.findById(id);
+            if (allowedUser.getUserId() == null) {
+                throw new NoContentFoundException("User Not Found!");
+            }
+            return new ResponseEntity<>(userService.update(user, id), HttpStatus.CREATED);
+
+        }
+        throw new UnAuthorizedException("User Not allowed for this operation!");
 
     }
 
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable UUID id) throws Exception {
-        userService.delete(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<?> delete(@PathVariable UUID id, @RequestHeader("Authorization") String token) throws Exception {
+        if (jwt.validateToken(token) && jwt.getUserFromToken(token).getUserId() != null) {
+            var allowedUser = userService.findById(id);
+            if (allowedUser.getUserId() == null) {
+                throw new NoContentFoundException("User Not Found!");
+            }
+            userService.delete(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+
+        }
+        throw new UnAuthorizedException("User Not allowed for this operation!");
+
     }
 
 }
+
+
